src/main/java/com/steakhouse/service/UserService.java
package com.steakhouse.service;

import com.steakhouse.exception.ResourceNotFoundException;
import com.steakhouse.model.User;
import com.steakhouse.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

import java.util.List;


@Service
public class UserService {

    @Autowired
    private UserRepository userRepository;

    @Autowired
    private PasswordEncoder passwordEncoder;

    // O‘z profilini yangilash
    public User updateUserByUsername(String username, User userDetails) {
        User user = userRepository.findByUsername(username)
                .orElseThrow(() -> new ResourceNotFoundException("User not found with username " + username));
        user.setUsername(userDetails.getUsername());
        user.setEmail(userDetails.getEmail());
        if (userDetails.getPassword() != null && !userDetails.getPassword().isEmpty()) {
            user.setPasswordHash(passwordEncoder.encode(userDetails.getPassword()));
        }
        return userRepository.save(user);
    }

    // O‘z profilini o‘chirish
    public void deleteUserByUsername(String username) {
        User user = userRepository.findByUsername(username)
                .orElseThrow(() -> new ResourceNotFoundException("User not found with username " + username));
        userRepository.delete(user);
    }

    public User registerUser(User user) {
        user.setPasswordHash(passwordEncoder.encode(user.getPassword()));
        user.setRole("ROLE_CUSTOMER");
        return userRepository.save(user);
    }

    public User findByUsername(String username) {
        return userRepository.findByUsername(username).orElse(null);
    }

    public User updateUser(Long id, User userDetails) {
        User user = userRepository.findById(id)
                .orElseThrow(() -> new ResourceNotFoundException("User not found with id " + id));
        user.setUsername(userDetails.getUsername());
        user.setEmail(userDetails.getEmail());
        user.setPasswordHash(passwordEncoder.encode(userDetails.getPassword())); // Parolni xavfsiz tarzda yangilash
        return userRepository.save(user);
    }

    public void deleteUser(Long id) {
        User user = userRepository.findById(id)
                .orElseThrow(() -> new ResourceNotFoundException("User not found with id " + id));
        userRepository.delete(user);
    }

    public List<User> getAllUsers() {
        return userRepository.findAll();
    }


}
